Support for JSONSchemaAlias method
